requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
beautifulsoup4
pydub
